% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_platform_velocity}
\alias{estimate_platform_velocity}
\title{Estimate platform velocity (and optionally u_x,u_y,h) from data}
\usage{
estimate_platform_velocity(data, known_u_x = NULL, init = c(4, 10, 10, 1))
}
\arguments{
\item{data}{data.frame with columns t, x, y}

\item{known_u_x}{numeric or NULL. If provided, estimate v = slope - known_u_x}

\item{init}{numeric initial guesses for optim when joint estimation: c(u_x, u_y, v, h)}
}
\value{
list with estimates and model object
}
\description{
If u_x is known, a simple linear regression on x ~ t recovers slope = u_x + v.
If u_x unknown, do a joint nonlinear least squares (optim) on x and y.
}
