% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{galilean_matrix}
\alias{galilean_matrix}
\title{Return the 2x2 Galilean matrix for x,t (1D)}
\usage{
galilean_matrix(v)
}
\arguments{
\item{v}{numeric scalar (velocity)}
}
\value{
2x2 matrix
}
\description{
Return the 2x2 Galilean matrix for x,t (1D)
}
