% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{galilean_transform}
\alias{galilean_transform}
\title{Apply a 1D Galilean boost along x}
\usage{
galilean_transform(pos, t, v)
}
\arguments{
\item{pos}{data.frame or matrix with columns x,y,z}

\item{t}{numeric vector of times (same length as rows of pos)}

\item{v}{numeric scalar (velocity along x)}
}
\value{
data.frame with x,y,z,t transformed (to the other frame)
}
\description{
Apply a 1D Galilean boost along x
}
\examples{
pos <- data.frame(x = c(0,1), y = c(0,0), z = c(0,0))
galilean_transform(pos, t = c(0,1), v = 5)
}
