% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galileantools.R
\docType{package}
\name{galileantools-package}
\alias{galileantools}
\alias{galileantools-package}
\title{galileantools: Tools for Galilean transformations and projectile simulations}
\description{
The galileantools package provides functions to:
\itemize{
\item Apply Galilean transformations
\item Simulate projectiles from moving platforms
\item Estimate platform velocity and projectile initial conditions
\item Visualize trajectories and diagnostics
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ourbani237/galileantools}
  \item Report bugs at \url{https://github.com/Ourbani237/galileantools/issues}
}

}
\author{
\strong{Maintainer}: Ibrahima OURBANI DOUBLA MAMOUDOU \email{ibrahimadoubla21@gmail.com}

}
