% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numerical_acceleration}
\alias{numerical_acceleration}
\title{Compute numerical acceleration from data (central differences)}
\usage{
numerical_acceleration(data)
}
\arguments{
\item{data}{data.frame with t,x,y}
}
\value{
data.frame with t, ax, ay
}
\description{
Compute numerical acceleration from data (central differences)
}
