% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot trajectories: observed noisy ground data and model fit if provided}
\usage{
plot_trajectories(data, model_par = NULL)
}
\arguments{
\item{data}{data.frame with t,x,y}

\item{model_par}{optional: named vector c(u_x,u_y,v,h) for model overlay}
}
\description{
Plot trajectories: observed noisy ground data and model fit if provided
}
